画像エンコーダをファイル分割
